% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ada.R
\name{get_dataset_ids}
\alias{get_dataset_ids}
\title{Get the IDs}
\usage{
get_dataset_ids()
}
\value{
A tibble with one column 'id' containing the IDs of the data.
}
\description{
Download the IDs of the available data.
}
