% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ada.R
\name{get_dataset_info}
\alias{get_dataset_info}
\title{Get dataset info with ID}
\usage{
get_dataset_info(id)
}
\arguments{
\item{id}{The ID of the data.}
}
\value{
A list with the following fields:
\describe{
  \item{\code{organization}}{The name of the organization
   that has released the data.}
  \item{\code{license}}{The license of the data.}
  \item{\code{datasets}}{If available, a list of the datasets.
   Contains fields name, format and url, the last of which points to the
   url of the dataset. Otherwise \code{NULL}.'}
  \item{\code{notes}}{Notes related to the data. Most likely in finnish.}
}
}
\description{
Download information related to the data.
}
